+++ services/routes/userRoutes/merchantRoute.js
@@ -6,6 +6,10 @@ const MerchantModel = require("../../models/User/MerchantModel");
 // get all
 router.get('/', async (req, res) => {
     try {
+        if (req.auth.admin === 0) {
+            res.status(401).send("Not authorised");
+            return;
+        }
         var merchants = await MerchantModel.find({}).exec();
         res.send(merchants);
     } catch (error) {
@@ -16,6 +20,10 @@ router.get('/', async (req, res) => {
 // get one
 router.get('/:id', async (req, res) => {
     try {
+        if (req.auth.isMerchant !== true && req.auth.admin === 0 || req.params.id !== req.auth.userid) {
+            res.status(401).send("Not authorised");
+            return;
+        }
         var merchant = await MerchantModel.findById(req.params.id).exec();
         res.send(merchant);
     } catch (error) {
@@ -37,6 +45,10 @@ router.post("/", async (req, res) => {
 // update existing
 router.put("/:id", async (req, res) => {
     try {
+        if (req.auth.isMerchant !== true && req.auth.admin === 0 || req.params.id !== req.auth.userid) {
+            res.status(401).send("Not authorised");
+            return;
+        }
         var merchant = await MerchantModel.findById(req.params.id).exec();
         merchant.set(req.body);
         var result = await merchant.save();
@@ -49,6 +61,10 @@ router.put("/:id", async (req, res) => {
 // delete one
 router.delete("/:id", async (req, res) => {
     try {
+        if (req.auth.admin === 0) {
+            res.status(401).send("Not authorised");
+            return;
+        }
         var result = await MerchantModel.deleteOne({ _id: req.params.id }).exec();
         res.send(result);
     } catch (error) {
